RevisionManager now accepts default adapter class on init
When constructing a revision manager you can now set a default adapter
class to use when registering models. If unset the usual VersionAdapter
class is used. The adapter can still be overridden when calling the
`register` method.

This makes it easy to use a custom adapter for all model registrations
without needing to provide the adapter with every `register` call.