Add support for "strict" revert/restore in Django admin
If the new optional Django setting REVERSION_ADMIN_STRICT_REVERT is
truthy, when reverting/restoring version data via the built-in Django
admin integration the exact prior version will be restored, and any
modifications made in the submitted admin form data will be ignored.

This change makes it possible to safely revert complex object graphs
without needing to guarantee that *every* piece of data in the prior
version is also properly represented in a single admin form.