When making a change during revert, create two revisions
When you want to revert to an older revision, you get a regular change
form with the older values pre-filled. If you make changes to the form
before saving, the new version saved will be different from the older
version you are reverting to, though the log message still only says
"Reverted to previous version".

This commit changes the reverting code to create a separate revision for
the revert, with the "Reverted to previous version" message. The
subsequent save of the form happens in a second revision.

This change is not complete: Currently the second revision is always
saved, regardless of whether any changes were made.