Improved performance of `createinitialrevisions` management command
Rather than retrieving the list of PKs that needs revisions created for, and then pulling out the actual objects in separate queries using `in_bulk`, we drop it down to a single `count()` query (to get the total number) and an `iterator()` query. This improves the performance (and possibly memory usage) on postgres databases when server side cursors are enabled (which they are by default on postgres databases). For databases where server side cursors are not enabled, we revert to the original code as `iterator()` is memory intensive (at the database driver level, see Django documentation on `iterator()` for further details).